Add files via upload
This sample ebook when loaded on Kobo or Nook ereaders will display the first file only (the Prologue).  The ebook displays in Calibre's reader normally, but in ereaders (tested in Kobo, Nook, & Readman) any chapter > the first file only shows a blank page. Turning to the next page causes the ereader to jump several pages (to the next chapter apparently), and then again only displays a blank page.